[PIL-2253] - update class names and more URLs
