+++ app/controllers/registration/ContactUPEByPhoneController.scala
@@ -19,40 +19,40 @@ package controllers.registration
 import config.FrontendAppConfig
 import connectors.UserAnswersConnectors
 import controllers.actions.{DataRequiredAction, DataRetrievalAction, IdentifierAction}
-import forms.CaptureTelephoneDetailsFormProvider
+import forms.ContactUPEByTelephoneFormProvider
 import models.Mode
 import navigation.UltimateParentNavigator
-import pages.{UpeCapturePhonePage, UpeContactNamePage, UpePhonePreferencePage}
+import pages.{UpeContactEmailPage, UpeContactNamePage, UpePhonePreferencePage}
 import play.api.i18n.I18nSupport
 import play.api.libs.json.Format.GenericFormat
 import play.api.libs.json.Json
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-import views.html.registrationview.CaptureTelephoneDetailsView
+import views.html.registrationview.ContactUPEByTelephoneView
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
-class CaptureTelephoneDetailsController @Inject() (
+class ContactUPEByPhoneController @Inject()(
   val userAnswersConnectors: UserAnswersConnectors,
   identify:                  IdentifierAction,
   getData:                   DataRetrievalAction,
   requireData:               DataRequiredAction,
-  formProvider:              CaptureTelephoneDetailsFormProvider,
+  formProvider:              ContactUPEByTelephoneFormProvider,
   val controllerComponents:  MessagesControllerComponents,
   navigator:                 UltimateParentNavigator,
-  view:                      CaptureTelephoneDetailsView
+  view:                      ContactUPEByTelephoneView
 )(implicit ec:               ExecutionContext, appConfig: FrontendAppConfig)
     extends FrontendBaseController
     with I18nSupport {
 
   def onPageLoad(mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData) { implicit request =>
     (for {
-      _           <- request.userAnswers.get(UpePhonePreferencePage)
+      _           <- request.userAnswers.get(UpeContactEmailPage)
       contactName <- request.userAnswers.get(UpeContactNamePage)
     } yield {
       val form = formProvider(contactName)
-      val preparedForm = request.userAnswers.get(UpeCapturePhonePage) match {
+      val preparedForm = request.userAnswers.get(UpePhonePreferencePage) match {
         case None        => form
         case Some(value) => form.fill(value)
       }
@@ -60,6 +60,7 @@ class CaptureTelephoneDetailsController @Inject() (
       Ok(view(preparedForm, mode, contactName))
     })
       .getOrElse(Redirect(controllers.routes.JourneyRecoveryController.onPageLoad()))
+
   }
 
   def onSubmit(mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData).async { implicit request =>
@@ -70,16 +71,15 @@ class CaptureTelephoneDetailsController @Inject() (
           .bindFromRequest()
           .fold(
             formWithErrors => Future.successful(BadRequest(view(formWithErrors, mode, contactName))),
-            value =>
+            nominated =>
               for {
                 updatedAnswers <-
-                  Future.fromTry(request.userAnswers.set(UpeCapturePhonePage, value))
+                  Future.fromTry(request.userAnswers.set(UpePhonePreferencePage, nominated))
                 _ <- userAnswersConnectors.save(updatedAnswers.id, Json.toJson(updatedAnswers.data))
-              } yield Redirect(navigator.nextPage(UpeCapturePhonePage, mode, updatedAnswers))
+              } yield Redirect(navigator.nextPage(UpePhonePreferencePage, mode, updatedAnswers))
           )
       }
       .getOrElse(Future.successful(Redirect(controllers.routes.JourneyRecoveryController.onPageLoad())))
-
   }
 
 }
